refactor: ChallengeItem 리팩터링
